requirements.txt
# django libs
Django == 3.2.11
uWSGI==2.0.19.1
whitenoise == 5.3.0
django-auth-ldap==3.0.0
django-user-agents == 0.4.0
django-filter == 21.1
django-cache-memoize == 0.1.10
django-ses == 2.3.1
django-storages == 1.12.3
django-celery-beat == 2.2.1
django-celery-results == 2.2.0
# django rest framework libs
djangorestframework == 3.13.1
drf-flex-fields == 0.9.6
djangorestframework-filters == 1.0.0.dev2
drf-spectacular == 0.21.1

git+https://github.com/certego/certego-saas.git@7a464e15813125ec33c49d5fcf3b75e13f191a09#egg=certego_saas
git+https://github.com/jazzband/django-silk.git#egg=django-silk

# infra
requests==2.26.0
boto3==1.20.6
celery==5.2.2
dataclasses==0.6
flower==1.0.0

# external integrations

# plugins
GitPython==3.1.18
checkdmarc==4.4.1
darksearch==2.1.2
dnspython==2.1.0
dnstwist==20211204
ppdeep>=20200505
easygui==0.98.2
google>=3.0.0
google-cloud-webrisk==1.6.0
intezer-sdk==1.6
maxminddb==2.2.0
mwdblib==3.4.0
oletools==0.60
OTXv2==1.5.12
peepdf-fork==0.4.3
pdfid==1.0.4
pefile-fork==2020.12.16
psycopg2-binary==2.9.1
pycti==5.1.3
pydeep==0.4
PyExifTool==0.4.9
pyhashlookup==1.1.1
pymisp==2.4.144
pypdns==1.5.2
pypssl==2.2
pysafebrowsing==0.1.1
PySocks==1.7.1
python-dotenv==0.19.2
quark-engine==22.1.1
speakeasy-emulator==1.5.9
yara-python==4.1.0
pydragonfly==0.0.3
XLMMacroDeobfuscator[secure]==0.2.3
# this is required because XLMMacroDeobfuscator does not pin xlrd2 and pyxlsb2 packages
pyxlsb2==0.0.8
xlrd2==1.3.4
